+++ CHANGES.md
@@ -3,3 +3,4 @@
 * add `class Loader` to gather plugins/themes with `Update URI` header data and run through Git Updater Lite
 * add checks to ensure correct data in `Update URI` header, in GUL
 * rebrand as lite-updater
+* re-write in FAIR style